requirements.txt
langchain_core
langchain_community
streamlit
python-dotenv
pypdf
